Created the actual Diner
Created the diner and added a little bit of light to the scene. We then just added a bit of lights to the scene to give it a bit of a pop. We are cool as hell. Just kidding but that is all and now phase 2 is to create the physics then the AI for the people.