Input Pressense and Hand Pressence
Firstly, I just added a feature that no matter the headset it will show the proper controller for your headset. Secondly, I added a feature that there are hands instead of controllers because I thought that just looked nicer.
Thirdly, Fixed the customer class and spawned so that the customers will pick a set when they are spawned and leave the door when they want to leave.