Added some UI
I added a bit UI to better be able to tell the Score, How much time to have and how many customers are left. Created the foundation of the object pooling but did not connect it to anything so it doesnt have purpose yet. The way to process the order is still broken but will fix after I get the rest working. changed the café again to better fit the aesthic and finally added a door to the café.